example/dqn/requirements.txt
gym==0.21.0
mindspore-gpu==1.8.1
numpy==1.21.6
onnx==1.12.0
onnxruntime==1.12.1
onnxruntime-gpu==1.12.1
mindspore_rl==0.7.0
importlib-metadata==4.12.0
